+++ lib/src/widgets/solid_security_key_manager_dialogs.dart
@@ -33,7 +33,7 @@ import 'package:flutter/material.dart';
 import 'package:solidpod/solidpod.dart'
     show SolidFunctionCallStatus, changeKeyPopup, getEncKeyPath, readPod;
 
-import 'package:solidui/src/widgets/solid_security_key_manager_helpers.dart';
+import 'package:solidui/src/widgets/solid_security_key_ui_helpers.dart';
 import 'package:solidui/src/widgets/solid_security_key_view.dart';
 
 /// Dialog management for Security Key Manager.
@@ -51,7 +51,7 @@ class SolidSecurityKeyManagerDialogs {
       await onKeyChanged();
     } catch (e) {
       if (context.mounted) {
-        SolidSecurityKeyManagerHelpers.showErrorSnackBar(context, e.toString());
+        SecurityKeyUIHelpers.showErrorSnackBar(context, e.toString());
       }
     }
   }
@@ -66,12 +66,15 @@ class SolidSecurityKeyManagerDialogs {
     Future<bool> Function(String key, String confirmKey)
         handleSubmissionFunction,
   ) async {
-    await SolidSecurityKeyManagerHelpers.showNewKeyDialog(
-      context,
-      keyController,
-      confirmKeyController,
-      onKeyChanged,
-      handleSubmissionFunction,
+    return showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (context) => _SetKeyDialog(
+        keyController: keyController,
+        confirmKeyController: confirmKeyController,
+        onKeyChanged: onKeyChanged,
+        handleSubmissionFunction: handleSubmissionFunction,
+      ),
     );
   }
 
@@ -82,7 +85,7 @@ class SolidSecurityKeyManagerDialogs {
     Future<void> Function() checkKeyStatus,
     Future<void> Function(BuildContext) showKeyInputDialog,
   ) async {
-    await SolidSecurityKeyManagerHelpers.showErrorDialog(
+    await SecurityKeyUIHelpers.showErrorDialog(
       context,
       'Security Key File Not Found',
       'The security key file could not be found. Would you like to set a new '
@@ -149,7 +152,7 @@ class SolidSecurityKeyManagerDialogs {
       if (!context.mounted) return;
 
       if (fileContent == SolidFunctionCallStatus.notLoggedIn.toString()) {
-        await SolidSecurityKeyManagerHelpers.showErrorDialog(
+        await SecurityKeyUIHelpers.showErrorDialog(
           context,
           'Not Logged In',
           'You must be logged in to view security keys.',
@@ -169,7 +172,7 @@ class SolidSecurityKeyManagerDialogs {
           ),
         );
       } else {
-        await SolidSecurityKeyManagerHelpers.showErrorDialog(
+        await SecurityKeyUIHelpers.showErrorDialog(
           context,
           'Empty Key File',
           'The security key file exists but appears to be empty.',
@@ -178,7 +181,7 @@ class SolidSecurityKeyManagerDialogs {
     } catch (e) {
       debugPrint('Exception reading security key: $e');
       if (context.mounted) {
-        await SolidSecurityKeyManagerHelpers.showErrorDialog(
+        await SecurityKeyUIHelpers.showErrorDialog(
           context,
           'Error Reading Key',
           e.toString(),
@@ -187,5 +190,189 @@ class SolidSecurityKeyManagerDialogs {
     } finally {
       setLoading(false);
     }
+  }
+}
+
+/// Dialog for setting a new security key with loading state.
+
+class _SetKeyDialog extends StatefulWidget {
+  final TextEditingController keyController;
+  final TextEditingController confirmKeyController;
+  final Future<void> Function() onKeyChanged;
+  final Future<bool> Function(String key, String confirmKey)
+      handleSubmissionFunction;
+
+  const _SetKeyDialog({
+    required this.keyController,
+    required this.confirmKeyController,
+    required this.onKeyChanged,
+    required this.handleSubmissionFunction,
+  });
+
+  @override
+  State<_SetKeyDialog> createState() => _SetKeyDialogState();
+}
+
+class _SetKeyDialogState extends State<_SetKeyDialog> {
+  bool _isLoading = false;
+  bool _obscureKey = true;
+  bool _obscureConfirmKey = true;
+
+  Future<void> _handleSetKey() async {
+    if (widget.keyController.text != widget.confirmKeyController.text) {
+      SecurityKeyUIHelpers.showErrorSnackBar(
+        context,
+        'Keys do not match',
+      );
+      return;
+    }
+    if (widget.keyController.text.length < 6) {
+      SecurityKeyUIHelpers.showErrorSnackBar(
+        context,
+        'Key must be at least 6 characters',
+      );
+      return;
+    }
+
+    // Show loading state
+    setState(() {
+      _isLoading = true;
+    });
+
+    try {
+      final success = await widget.handleSubmissionFunction(
+        widget.keyController.text,
+        widget.confirmKeyController.text,
+      );
+
+      if (!mounted) return;
+
+      if (success) {
+        // Update the key status first.
+
+        await widget.onKeyChanged();
+
+        if (!mounted) return;
+
+        // Close the Set Key dialog.
+
+        Navigator.of(context).pop();
+
+        if (!mounted) return;
+
+        // Close the Security Key Manager dialog.
+
+        Navigator.of(context).pop();
+
+        // Show success dialog.
+
+        await SecurityKeyUIHelpers.showErrorDialog(
+          context,
+          'Success',
+          'Security key has been set successfully.',
+        );
+      } else {
+        // Only hide loading if operation failed (to allow retry).
+
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    } catch (e) {
+      if (!mounted) return;
+
+      setState(() {
+        _isLoading = false;
+      });
+
+      SecurityKeyUIHelpers.showErrorSnackBar(
+        context,
+        'Failed to set key: $e',
+      );
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return PopScope(
+      canPop: !_isLoading,
+      child: AlertDialog(
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+        title: const Text(
+          'Set Security Key',
+          style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+        ),
+        content: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            if (!_isLoading) ...[
+              TextField(
+                controller: widget.keyController,
+                decoration: SecurityKeyUIHelpers.getInputDecoration(
+                  'Enter Security Key',
+                  ThemeData(),
+                ).copyWith(
+                  suffixIcon: IconButton(
+                    icon: Icon(
+                      _obscureKey ? Icons.visibility : Icons.visibility_off,
+                    ),
+                    onPressed: () {
+                      setState(() {
+                        _obscureKey = !_obscureKey;
+                      });
+                    },
+                  ),
+                ),
+                obscureText: _obscureKey,
+              ),
+              const SizedBox(height: 16),
+              TextField(
+                controller: widget.confirmKeyController,
+                decoration: SecurityKeyUIHelpers.getInputDecoration(
+                  'Confirm Security Key',
+                  ThemeData(),
+                ).copyWith(
+                  suffixIcon: IconButton(
+                    icon: Icon(
+                      _obscureConfirmKey
+                          ? Icons.visibility
+                          : Icons.visibility_off,
+                    ),
+                    onPressed: () {
+                      setState(() {
+                        _obscureConfirmKey = !_obscureConfirmKey;
+                      });
+                    },
+                  ),
+                ),
+                obscureText: _obscureConfirmKey,
+              ),
+            ] else ...[
+              const SizedBox(height: 20),
+              const CircularProgressIndicator(),
+              const SizedBox(height: 20),
+              const Text(
+                'Setting security key...',
+                style: TextStyle(fontSize: 14),
+              ),
+              const SizedBox(height: 20),
+            ],
+          ],
+        ),
+        actions: _isLoading
+            ? []
+            : [
+                TextButton(
+                  onPressed: () => Navigator.of(context).pop(),
+                  child: const Text('Cancel'),
+                ),
+                ElevatedButton(
+                  onPressed: _handleSetKey,
+                  style: SecurityKeyUIHelpers.getButtonStyle(ThemeData()),
+                  child: const Text('Set Key'),
+                ),
+              ],
+      ),
+    );
   }
 }
